refactor: useProfile 공통 훅 분리
