+++ src/features/user/hooks.js
@@ -1,28 +1,12 @@
-import {useState, useEffect} from 'react';
-import {fetchMyProfile, saveDesignerProfile, saveBusinessProfile, uploadProfileImage} from '@/api/user';
+import {saveDesignerProfile, saveBusinessProfile, uploadProfileImage} from '@/api/user';
 import {useStore} from '@/store/useStore';
-
-export const useProfile = () => {
-  const [profile, setProfile] = useState(null);
-  useEffect(() => {
-    (async () => {
-      try {
-        const data = await fetchMyProfile();
-        setProfile(data);
-      } catch {}
-    })();
-  }, []);
-  return {profile};
-};
 
 export const useSaveDesigner = () => {
   const save = async (designerData) => {
     try {
-      const result = await saveDesignerProfile(designerData);
-      return result;
+      return await saveDesignerProfile(designerData);
     } catch (err) {
-      const msg = err?.response?.data?.message || err?.message || '저장 실패';
-      alert(msg);
+      alert(err?.message || '저장 실패');
       return null;
     }
   };
@@ -32,11 +16,9 @@ export const useSaveDesigner = () => {
 export const useSaveBusiness = () => {
   const save = async (businessData) => {
     try {
-      const result = await saveBusinessProfile(businessData);
-      return result;
+      return await saveBusinessProfile(businessData);
     } catch (err) {
-      const msg = err?.response?.data?.message || err?.message || '저장 실패';
-      alert(msg);
+      alert(err?.message || '저장 실패');
       return null;
     }
   };
@@ -49,10 +31,9 @@ export const useUploadImage = () => {
   const upload = async (file) => {
     if (!file) return null;
     try {
-      const result = await uploadProfileImage(file, userType);
-      return result;
+      return await uploadProfileImage(file, userType);
     } catch (err) {
-      console.error('이미지 업로드 실패:', err);
+      alert(err?.message || '이미지 업로드 실패');
       return null;
     }
   };
